requirements.txt
torch 
torchvision 
fastapi 
uvicorn 
pillow 
python-multipart 
scikit-learn 
matplotlib